fix for batch processor spx fitting
When PyMca does batch fits on .spx files, it will interpret the SNIP background from the bruker software as its own spectrum, and fit both the actual spectrum and the background as though they were both spectra. When it does this, the csv file contains peak areas from the actual spectrum in column 0, and peak areas from the background in column 1.  the pymca spreadsheet reading cell has been modified to exclude data designated as column 1.